Move to a new home that doesn't conflict with Mesos-Util
