+++ src/test/scala/mesosphere/marathon/util/TimeoutTest.scala
@@ -1,4 +1,4 @@
-package mesosphere.util
+package mesosphere.marathon.util
 
 import mesosphere.AkkaUnitTest
